# Fix a regression in watch performance and reduce allocations in the GET path