#75509 identified that 6f9bf5fe98bcc3b436fea4d6dd345a1502d20778 caused a performance regression. The regression was due to unifying the watch code path with the short-running request path - we used to have a special "watch only" self-link setter that bypassed having to check whether the object was a list or not. That check `IsListType` was hugely expensive in terms of allocations caused allocations per object to jump from 4 (proto) or 30 (json) per watch event to over 150 for json.

Make `IsListType()` faster since it is used widely by removing the use of `%#v` and cache the type lookup.  That takes both the error return and reflection out for all of our types.

Second, bypass IsListType() for the special case of self link generation in the apiserver - only objects that identify as meta v1 List interfaces (has Items *and* has ListMeta) should have nested self-links set.

Profiles at this point revealed some other low hanging fruit in get that the subsequent commits address.

The metrics cleanUserAgent method can avoid allocating in the normal case.

The metrics cleanDryRun can avoid allocating in the normal case

Negotiation created a bunch of ephemeral objects.

```
# before
BenchmarkGet-12          	  100000	    119635 ns/op	   20110 B/op	     206 allocs/op
BenchmarkWatchHTTP-12    	  100000	     65761 ns/op	    7296 B/op	     139 allocs/op

# after
BenchmarkGet-12          	  100000	    108850 ns/op	   15942 B/op	     132 allocs/op
BenchmarkWatchHTTP-12    	  200000	     34133 ns/op	    1913 B/op	      30 allocs/op
```

Which restores watch to its previous levels and significantly improves GET (note that GC cost is not accounted in profile time and is significantly more expensive than the bump allocation used in go)

/kind bug

```release-note
NONE
```