@wojtek-t 2019-03-22
The failures seems legitimate:
```
no Items field in *unstructured.Unstructured
```