@smarterclayton 2019-03-22
Yeah, was a legit failure because metav1.Status implements ListMeta.  Just using the simpler cache, I may open a proposal requiring internal types that want to  be treated like a list behave as one.  This is one of our last reflection code paths in the code and I'd be happy to see it die.