@liggitt 2019-03-22
two comments on the cache, one question on the new typetype/typesubtype fields for out-of-tree callers. lgtm otherwise